requirements.txt
numpy
ortools
python-lifelib
scipy
